# specs/001-neon-migration/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo API
  description: RESTful API for Todo application with JWT authentication
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://todo-api-production.vercel.app/api/v1
    description: Production server
...
paths:
...
  /todos:
    get:
      tags: [todos]
      summary: List user's todos
      description: Get all active todos for authenticated user (paginated)
      operationId: listTodos
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/CompletedParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Todos retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'

    post:
      tags: [todos]
      summary: Create new todo
      description: Create a new todo for authenticated user
      operationId: createTodo
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoCreate'
            example:
              title: "Buy groceries"
              description: "Milk, eggs, bread"
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /todos/{todo_id}:
    get:
      tags: [todos]
      summary: Get todo by ID
      description: Retrieve a specific todo owned by authenticated user
      operationId: getTodo
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TodoIdParam'
      responses:
        '200':
          description: Todo retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    put:
      tags: [todos]
      summary: Update todo
      description: Update a specific todo owned by authenticated user
      operationId: updateTodo
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TodoIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoUpdate'
            example:
              title: "Buy groceries TODAY"
              completed: true
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [todos]
      summary: Delete todo (soft delete)
      description: Soft delete a specific todo owned by authenticated user
      operationId: deleteTodo
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TodoIdParam'
      responses:
        '204':
          description: Todo deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (expires in 15 minutes)
  parameters:
    TodoIdParam:
      name: todo_id
      in: path
      required: true
      description: UUID of the todo
      schema:
        type: string
        format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    PageParam:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    PageSizeParam:
      name: page_size
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
    CompletedParam:
      name: completed
      in: query
      description: Filter by completion status
      schema:
        type: boolean
      example: false
    SortByParam:
      name: sort_by
      in: query
      description: Field to sort by
      schema:
        type: string
        enum: [created_at, updated_at, title]
        default: created_at
      example: created_at
    SortOrderParam:
      name: sort_order
      in: query
      description: Sort order
      schema:
        type: string
        enum: [asc, desc]
        default: desc
      example: desc
  schemas:
...
    TodoCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Todo title
          example: "Buy groceries"
        description:
          type: string
          maxLength: 5000
          description: Todo description (optional)
          example: "Milk, eggs, bread"
    TodoUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Updated todo title
        description:
          type: string
          maxLength: 5000
          description: Updated todo description
        completed:
          type: boolean
          description: Todo completion status
    Todo:
      type: object
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Todo's unique identifier
        user_id:
          type: string
          format: uuid
          description: Owner's user ID
        title:
          type: string
          description: Todo title
        description:
          type: string
          nullable: true
          description: Todo description
        completed:
          type: boolean
          description: Todo completion status
        deleted_at:
          type: string
          format: date-time
          nullable: true
          description: Soft delete timestamp (null if active)
        created_at:
          type: string
          format: date-time
          description: Todo creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
    TodoListResponse:
      type: object
      required:
        - todos
        - total
        - page
        - page_size
      properties:
        todos:
          type: array
          items:
            $ref: '#/components/schemas/Todo'
        total:
          type: integer
          description: Total number of todos (across all pages)
        page:
          type: integer
          description: Current page number
        page_size:
          type: integer
          description: Number of items per page
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type/category
          example: "validation_error"
        message:
          type: string
          description: Human-readable error message
          example: "Email is required"
        details:
          type: object
          description: Additional error details (optional)
          additionalProperties: true
  responses:
    BadRequest:
      description: Bad request (invalid input)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "bad_request"
            message: "Passwords do not match"
            details: {}
    Unauthorized:
      description: Unauthorized (invalid or missing token)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "unauthorized"
            message: "Invalid or expired token"
            details: {}
    Forbidden:
      description: Forbidden (insufficient permissions)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "forbidden"
            message: "You don't have permission to access this resource"
            details: {}
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "not_found"
            message: "Todo not found"
            details: {}
...
    ValidationError:
      description: Validation error (invalid input format)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "validation_error"
            message: "Invalid input data"
            details:
              email: "Invalid email format"
              password: "Password must be at least 8 characters"